requirements.txt
astroid==2.3.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
emoji==0.5.4
future==0.18.2
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.7
pyTelegramBotAPI==3.7.1
python-dateutil==2.8.1
python-telegram-bot==12.7
requests==2.23.0
six==1.14.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2